inotify: Fix a few trivial -Wsign-conversion warnings
When building GIO with `-Wsign-conversion` enabled.

In all of these cases, the implicit sign conversion cast is correct, so
let’s make it explicit to indicate that we‘ve checked it.


Helps: #3405

+++ gio/inotify/inotify-kernel.c
@@ -144,7 +144,7 @@ ik_source_can_dispatch_now (InotifyKernelSource *iks,
   return 0 <= dispatch_time && dispatch_time <= now;
 }
 
-static gsize
+static size_t
 ik_source_read_some_events (InotifyKernelSource *iks,
                             gchar               *buffer,
                             gsize                buffer_len)
@@ -169,7 +169,7 @@ again:
   else if (result == 0)
     g_error ("inotify unexpectedly hit eof");
 
-  return result;
+  return (size_t) result;
 }
 
 static gchar *
